python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bahamut"
version = "0.1.0"
description = "Client library for the followthemoney statement data server."
readme = "README.md"
license = "MIT"
authors = [{ name = "OpenSanctions", email = "info@opensanctions.org" }]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
requires-python = ">= 3.11"
dependencies = [
    "followthemoney == 3.*",
    "nomenklatura == 3.*",
    "grpcio >= 1.70.0",
]

[project.urls]
Documentation = "https://github.com/opensanctions/bahamut/"
Repository = "https://github.com/opensanctions/bahamut.git"
Issues = "https://github.com/opensanctions/bahamut/issues"

[project.optional-dependencies]
dev = [
    "black==24.10.0",
    "isort==5.13.2",
    "wheel>=0.29.0",
    "twine",
    "build",
    "mypy",
    "flake8>=2.6.0",
    "pytest",
    "ruff>=0.4.0,<1.0.0",
    "pytest-cov",
    "grpcio-tools >= 1.70.0",
    "protoletariat",
    "coverage>=4.1",
    "types-setuptools",
]

[tool.hatch.build.targets.sdist]
only-include = ["bahamut", "README.md"]

[tool.hatch.build.targets.wheel]
only-include = ["bahamut", "README.md"]

[tool.distutils.bdist_wheel]
universal = true

[tool.coverage.run]
branch = true
